% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Funcion para el calculo del AUC y ROC}
\usage{
AUC(df)
}
\arguments{
\item{df}{será un data frame/matriz que contendrá al menos una columna de clase numeric y otra columna binaria de clase factor}
}
\value{
Una lista de clase numeric (el area y los pares de puntos del ROC)
}
\description{
Esta funcion calcula el area debajo de la curva (AUC) y los puntos de la curva ROC
}
\examples{
set.seed(123)
data_frame <- data.frame(
Variable1 = rnorm(100),  # 100 números aleatorios
VariableBinaria = factor(sample(c(0, 1), 100, replace = TRUE)))
AUC(data_frame)
}
