% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo_metricas.R
\name{calcular_metricas}
\alias{calcular_metricas}
\title{Funcion para el calculo de la varianza y AUC cuando la variable es numerica y la entropia si es categorica}
\usage{
calcular_metricas(df)
}
\arguments{
\item{df}{será un data frame que puede contener columnas de clase numeric y de clase factor}
}
\value{
Un data frame de clase numeric
}
\description{
Esta funcion calcula el area debajo de la curva (AUC) y la varianza cuando la variable es numerica y calcula la entropia si la variable es categorica
}
\details{
utiliza las funciones \code{\link{var_vector}}, \code{\link{AUC}} y \code{\link{entropy}} internamente
}
\examples{
data <- data.frame(Score = as.numeric(c(0.1, 0.4, 0.35, 0.6, 0.3, 0.75, 0.55, 0.8)),
Outcome = as.factor(c("Negative", "Negative", "Negative", "Positive", "Negative", "Positive", "Positive", "Positive")))
calcular_metricas(data)
}
