% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{cor_fun}
\alias{cor_fun}
\title{Funcion de la correlación de las columnas de un data frame/matriz}
\usage{
cor_fun(matriz_dat)
}
\arguments{
\item{matriz_dat}{será un data frame/matriz de clase numeric}
}
\value{
Una matriz de clase numeric (la matriz de correlaciones)
}
\description{
Esta funcion calcula la correlación de las columnas de un data frame/matriz
}
\details{
utiliza la funcion \code{\link{cov_fun}} internamente
}
\examples{
df <- data.frame(var1 = c(10, 20, 30, 40, 50),
var2 = c(5, 15, 25, 35, 45),
var3 = c(2, 4, 6, 8, 10))
cor_fun(df)
}
