% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculo_metricas.R
\name{cor_infomutua}
\alias{cor_infomutua}
\title{Funcion para el calculo de la correlacion entre pares cuando las variables son numericas y la informacion mutua si son categoricas}
\usage{
cor_infomutua(df)
}
\arguments{
\item{df}{será un data frame que puede contener columnas de clase numeric y de clase factor}
}
\value{
Un data frame de clase numeric
}
\description{
Esta funcion calcula la correlacion entre pares de variables cuando las variables son numericas y calcula la informacion mutua si las variables son categoricas.
Si una columna es numerica y otra categorica lo ignora y el resultado es NA
}
\details{
utiliza las funciones \code{\link{cor_fun}} y \code{\link{info_mutua}} internamente
}
\examples{
dataset <- data.frame(var_num1 = rnorm(100),                 # Variable numérica
var_num2 = rnorm(100, mean = 5),       # Otra variable numérica
var_cat1 = factor(sample(letters[1:3], 100, replace = TRUE)),  # Variable categórica con 3 categorías
var_cat2 = factor(sample(letters[1:4], 100, replace = TRUE))   # Variable categórica con 4 categorías)
cor_infomutua(dataset)
}
