% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{cov_fun}
\alias{cov_fun}
\title{Funcion de la covarianza de las columnas de un data frame/matriz}
\usage{
cov_fun(x)
}
\arguments{
\item{x}{será un data frame/matriz de clase numeric}
}
\value{
Una matriz de clase numeric (la matriz de varianzas-covarianzas)
}
\description{
Esta funcion calcula la covarianza de las columnas de un data frame/matriz
}
\examples{
df <- data.frame(var1 = c(10, 20, 30, 40, 50),
var2 = c(5, 15, 25, 35, 45),
var3 = c(2, 4, 6, 8, 10))
cov_fun(df)
}
