% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretize}
\alias{discretize}
\title{Funcion de discretizacion dado los puntos de corte}
\usage{
discretize(x, cut.points)
}
\arguments{
\item{x}{será un vector de clase numeric}

\item{cut.points}{es una lista de puntos de corte de clase numeric}
}
\value{
Un vector de clase factor
}
\description{
Esta funcion discretiza un vector numerico en intervalos utilizando los puntos de corte dados
}
\examples{
a <- c(100, 50, 60, 10, 80, 20)
puntoscorte<- list(32.5, 55.0, 77.5)
discretize(a, puntoscorte)
}
