% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizeEF}
\alias{discretizeEF}
\title{Funcion de discretizacion Equal Frecuency}
\usage{
discretizeEF(x, num.bins)
}
\arguments{
\item{x}{será un vector de clase numeric}

\item{num.bins}{es un numero que indica cuantos intervalos queremos de clase numeric}
}
\value{
Un vector de clase factor
}
\description{
Esta funcion discretiza un vector numerico en intervalos que contengan el mismo numero de elementos
}
\examples{
a <- c(100, 50, 60, 10, 80, 20)
discretizeEF(a, 4)
}
