% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizacion.R
\name{discretizeEW}
\alias{discretizeEW}
\title{Funcion de discretizacion Equal Width}
\usage{
discretizeEW(x, num.bins)
}
\arguments{
\item{x}{será un vector de clase numeric}

\item{num.bins}{es un numero que indica cuantos intervalos queremos de clase numeric}
}
\value{
Un vector de clase factor
}
\description{
Esta funcion discretiza un vector numerico en intervalos de longitud iguales
}
\examples{
a <- c(100, 50, 60, 10, 80, 20)
discretizeEW(a, 4)
}
