% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropia.R
\name{entropy}
\alias{entropy}
\title{Funcion para calcular la entropia}
\usage{
entropy(x)
}
\arguments{
\item{x}{será un vector de clase factor}
}
\value{
un valor de clase numeric
}
\description{
Esta funcion calcula la entropia de un vector de clase factor
}
\examples{
data_factor <- factor(c("A", "B", "A", "C", "B", "A", "D", "D", "C", "A"))
entropy(data_factor)
}
