% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropia.R
\name{entropy_df}
\alias{entropy_df}
\title{Funcion para calcular la entropia de cada columna de un data frame}
\usage{
entropy_df(x)
}
\arguments{
\item{x}{será un dataframe de clase factor}
}
\value{
un vector de clase numeric
}
\description{
Esta funcion calcula la entropia de cada columna de un data frame de clase factor
}
\details{
utiliza la funcion \code{\link{entropy}} internamente
}
\examples{
data_frame <- data.frame(Color = factor(c("Rojo", "Verde", "Azul", "Rojo", "Verde", "Verde")),
Forma = factor(c("Círculo", "Cuadrado", "Círculo", "Cuadrado", "Cuadrado", "Círculo")),
Tamaño = factor(c("Grande", "Pequeño", "Pequeño", "Grande", "Grande", "Pequeño")))
entropy_df(data_frame)
}
