% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{estandar}
\alias{estandar}
\title{Funcion de estandarizacion de un vector}
\usage{
estandar(v)
}
\arguments{
\item{v}{será un vector de clase numeric}
}
\value{
Un vector de clase numeric
}
\description{
Esta funcion estandariza un vector numerico
}
\examples{
vector <- c(10, 20, 30, 40, 50)
estandar(vector)
}
