% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{info_mutua}
\alias{info_mutua}
\title{Funcion de la informaci칩n mutua de las columnas de un data frame/matriz}
\usage{
info_mutua(x)
}
\arguments{
\item{x}{ser치 un data frame/matriz de clase factor}
}
\value{
Una matriz de clase numeric (la matriz de informaci칩n mutua)
}
\description{
Esta funcion calcula la informaci칩n mutua de las columnas de un data frame/matriz
}
\details{
utiliza la funcion \code{\link{entropy_df}} internamente para calcular las entropias de cada columna
}
\examples{
data <- data.frame(
Columna_A = as.factor(c("Rojo", "Rojo", "Azul", "Verde", "Rojo", "Azul", "Verde", "Verde")),
Columna_B = as.factor(c("Alto", "Medio", "Medio", "Bajo", "Alto", "Bajo", "Alto", "Medio")),
Columna_C = as.factor(c("Bajo", "Alto", "Medio", "Bajo", "Medio", "Bajo", "Alto", "Medio")))
info_mutua(data)
}
