% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{norm}
\alias{norm}
\title{Funcion de normalización de un vector}
\usage{
norm(v)
}
\arguments{
\item{v}{será un vector de clase numeric}
}
\value{
Un vector de clase numeric
}
\description{
Esta funcion normaliza un vector numerico
}
\examples{
vector <- c(10, 20, 30, 40, 50)
norm(vector)
}
