% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropia.R
\name{norm_entropy}
\alias{norm_entropy}
\title{Funcion para calcular la entropia normalizada}
\usage{
norm_entropy(x)
}
\arguments{
\item{x}{será un vector de clase factor}
}
\value{
un valor de clase numeric
}
\description{
Esta funcion calcula la entropia normalizada de un vector de clase factor
}
\details{
utiliza la función \code{\link{entropy}} internamente para calcular la entropia del vector
}
\examples{
data_factor <- factor(c("A", "B", "A", "C", "B", "A", "D", "D", "C", "A"))
norm_entropy(data_factor)
}
