% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_AUC}
\alias{plot_AUC}
\title{Funcion para visualizar la curva ROC}
\usage{
plot_AUC(x)
}
\arguments{
\item{x}{será un data frame con una columna lógica o binaria y las demás numéricas}
}
\value{
Un gráfico de la curva ROC
}
\description{
Esta funcion calcula la Tasa de Falsos positivos (FPR) y la Tasa de Verdaderos Positivos (TPR)
para cada valor de corte y los grafica junto con la curva del random chance
}
\details{
Utiliza la funcion \code{\link{AUC}} dentro de la funcion para la visualizacion [mira la documentación correspondiente]
}
\examples{
# Paso 1: Crear el DataFrame con valores específicos
#Columna numérica
numerica <- c(2.5, 3.0, 4.5, 5.5, 6.0, 7.2, 8.1, 9.0, 10.0, 11.5)
#Columna lógica (TRUE si el valor es mayor que 6, FALSE en caso contrario)
logica <- c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, TRUE)
#Crear el DataFrame
df <- data.frame(numerica = numerica, logica = logica)
# Paso 2: Definir un vector de valores de corte
# En este caso, usaremos una secuencia de valores de corte
valores_corte <- seq(min(df$numerica), max(df$numerica, na.rm = TRUE), length.out = 200)
plot_AUC(df, valores_corte)
}
