% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cor}
\alias{plot_cor}
\title{Funcion para visualizar el heatmap de una matriz de correlaciones}
\usage{
plot_cor(x)
}
\arguments{
\item{x}{será un data frame / matriz con variables numéricas}
}
\value{
Un mapa de calor para la matriz de correlaciones
}
\description{
Esta funcion calcula la matriz de correlaciones para luego poder hacer un mapa de calor
}
\details{
Utiliza la funcion \code{\link{cor_fun}} dentro de la funcion para la visualizacion [mira la documentación correspondiente]
}
\examples{
set.seed(123)
datos_ejemplo <- data.frame(
Var1 = rnorm(100),
Var2 = rnorm(100),
 Var3 = rnorm(100),
 Var4 = rnorm(100))
 plot_cor(datos_ejemplo)
}
