% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_entropy}
\alias{plot_entropy}
\title{Funcion para visualizar la entropia normalizada de las variables de un data set}
\usage{
plot_entropy(df)
}
\arguments{
\item{df}{ser치 un data frame con variables categ칩ricas de tipo factor}
}
\value{
Un gr치fico de barras para la entropia normalizada de cada variable
}
\description{
Esta funcion calcula la entropia normalizada para luego poder hacer un grafico de barras con cada variable
}
\details{
Utiliza la funcion \code{\link{norm_entropy_df}} dentro de la funcion para la visualizacion [mira la documentaci칩n correspondiente]
}
\examples{
data <- data.frame(Columna_A = as.factor(c("Rojo", "Rojo", "Azul", "Verde", "Rojo", "Azul", "Verde", "Verde")),
Columna_B = as.factor(c("Alto", "Medio", "Medio", "Bajo", "Alto", "Bajo", "Alto", "Medio")),
Columna_C = as.factor(c("Bajo", "Alto", "Medio", "Bajo", "Medio", "Bajo", "Alto", "Medio")))
plot_entropy(data)
}
