% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_infomutua}
\alias{plot_infomutua}
\title{Funcion para visualizar la informacion mutua de un data set}
\usage{
plot_infomutua(x)
}
\arguments{
\item{x}{ser치 un data frame / matriz con variables categ칩ricas de tipo factor}
}
\value{
Un mapa de calor para la matriz de informacion mutua
}
\description{
Esta funcion calcula la matriz de la informaci칩n mutua para luego poder hacer un mapa de calor
}
\details{
Utiliza la funcion \code{\link{info_mutua}} dentro de la funcion para la visualizacion [mira la documentaci칩n correspondiente]
}
\examples{
data <- data.frame(Columna_A = as.factor(c("Rojo", "Rojo", "Azul", "Verde", "Rojo", "Azul", "Verde", "Verde")),
Columna_B = as.factor(c("Alto", "Medio", "Medio", "Bajo", "Alto", "Bajo", "Alto", "Medio")),
Columna_C = as.factor(c("Bajo", "Alto", "Medio", "Bajo", "Medio", "Bajo", "Alto", "Medio")))
plot_infomutua(data)
}
