% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Funcion para encontrar y eliminar filas duplicadas en un data frame}
\usage{
remove_duplicates(data)
}
\arguments{
\item{data}{será un dataframe de cualquier clase}
}
\value{
una dataframe de la misma clase que hemos proporcionado
}
\description{
Esta funcion encuentra y elimina filas del dataset que están repetidas
}
\examples{
data_with_duplicates <- data.frame(
ID = c(1, 2, 2, 3, 4, 4, 5),
Value = c("A", "B", "B", "C", "D", "D", "E"))
remove_duplicates(data_with_duplicates)
}
