% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{split_data}
\alias{split_data}
\title{Funcion para dividir dataset en entrenamiento y prueba}
\usage{
split_data(data, train_ratio = 0.7, seed = NULL)
}
\arguments{
\item{data}{será un dataframe/matriz de clase numeric}

\item{train_ratio}{indica el porcentaje de los datos originales que iran para el set de entrenamiento el default es 0.7}

\item{seed}{valor de tipo numeric para reproducibilidad}
}
\value{
una lista de clase numeric
}
\description{
Esta funcion divide el dataset en dos uno para el entrenamiento y otro para prueba
}
\examples{
data <- data.frame(
ID = 1:10,
Value = c(5, 3, 8, 2, 6, 9, 1, 7, 4, 10))
split_data(simple_data, train_ratio = 0.7, seed = 123)
}
