% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{var}
\alias{var}
\title{Funcion de la varianza de las columnas de un data frame/matriz}
\usage{
var(x)
}
\arguments{
\item{x}{será un data frame/matriz de clase numeric}
}
\value{
Un vector de clase numeric
}
\description{
Esta funcion calcula la varianza de las columnas de un data frame/matriz
}
\examples{
df <- data.frame(columna1 = c(10, 20, 30, 40, 50),
columna2 = c(5, 15, 25, 35, 45),
columna3 = c(2, 4, 6, 8, 10))
var(df)
}
