% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_analisis.R
\name{var_vector}
\alias{var_vector}
\title{Funcion de la varianza de un vector}
\usage{
var_vector(x)
}
\arguments{
\item{x}{será un vector de clase numeric}
}
\value{
Un valor de clase numeric
}
\description{
Esta funcion calcula la varianza de un vector numerico
}
\examples{
vector <- c(10, 20, 30, 40, 50)
var_vector(vector)
}
